almost end of project
Because the build of the game didn’t wanted to load the game after the cutscenes, and I didn’t knew how to fix this, I decided to leave the cutscene’s out so you could at least play the game. (What there is to play, because it’s not much).